Fitness_tracker/requirements.txt
# Core scientific
numpy>=1.26,<1.27
scipy>=1.15,<1.16
scikit-learn>=1.7,<1.8
matplotlib>=3.10,<3.11
opencv-python>=4.9.0.80,<4.10
opencv-contrib-python>=4.9.0.80,<4.10
contourpy>=1.3,<1.4
cycler>=0.12,<0.13
kiwisolver>=1.4,<1.5

# ML / AI
torch==2.8.0
torchvision==0.23.0
jax>=0.6.2,<0.7
jaxlib>=0.6.2,<0.7
mediapipe>=0.10,<0.11
ml-dtypes>=0.5,<0.6
opt-einsum>=3.4,<3.5
flatbuffers>=25.9,<26.0
sentencepiece>=0.2,<0.3

# Web / API
fastapi>=0.119,<0.120
uvicorn>=0.37,<0.38
python-jose>=3.5,<3.6
python-multipart>=0.0.20,<0.0.21
pydantic>=2.12,<2.13
pydantic-core>=2.41,<2.42
starlette>=0.48,<0.49
anyio>=4.11,<4.12
h11>=0.16,<0.17

# Utilities
absl-py>=2.3,<2.4
annotated-types>=0.7,<0.8
attrs>=25.4,<25.5
beautifulsoup4>=4.14,<4.15
certifi>=2025.10,<2025.11
cffi>=2.0,<2.1
charset-normalizer>=3.4,<3.5
click>=8.3,<8.4
colorama>=0.4,<0.5
fonttools>=4.60,<4.61
google>=3.0,<3.1
idna>=3.11,<3.12
packaging>=25.0,<25.1
passlib>=1.7,<1.8
pillow>=11.3,<11.4
protobuf>=4.25,<4.26
pycparser>=2.23,<2.24
pyparsing>=3.2,<3.3
python-dateutil>=2.9,<2.10
pyttsx3>=2.99,<3.0
requests>=2.32,<2.33
six>=1.17,<1.18
sniffio>=1.3,<1.4
sounddevice>=0.5,<0.6
soupsieve>=2.8,<2.9
typing-extensions>=4.15,<4.16
typing-inspection>=0.4,<0.5
urllib3>=2.5,<2.6

